+++ changelog.md
@@ -37,6 +37,7 @@
 * Quantity field was shown though the add to cart button was disabled
 * #260 Delivery times translations are not getting applied
 * robots.txt: localizable disallow paths did not contain a trailing slash
+* #296 Fix price adjustment of product variant combinations
 
 ###Improvements###
 * #171: select2 Dropdown too short in OPC